requirements.txt
Flask~=3.1.0
Flask-RESTful
opencv-python~=4.10.0.84
numpy~=2.1.3
firebase-admin~=6.6.0
Pillow~=11.0.0

rasterio~=1.4.2
ultralytics~=8.3.39
torch~=2.5.1
Flask-Cors~=5.0.0
pandas~=2.2.3
flet~=0.25.1
np~=1.0.2
requests~=2.32.3